fix dev setup and remove docker/vscode workspace
