--- .vscode/settings.json
@@ -1,18 +1,0 @@
-{
-    "python.defaultInterpreterPath": "frappe-bench/env/bin/python",
-    "debug.node.autoAttach": "disabled",
-    "sqltools.connections": [
-        {
-            "mysqlOptions": {
-                "authProtocol": "default"
-            },
-            "previewLimit": 50,
-            "server": "mariadb",
-            "port": 3306,
-            "driver": "MariaDB",
-            "name": "MariaDB",
-            "username": "root",
-            "password": "123"
-        }
-    ]
-}
