--- docker/init.sh
@@ -1,39 +1,0 @@
-#!bin/bash
-
-if [ -d "/home/frappe/frappe-bench/apps/frappe" ]; then
-    echo "Bench already exists, skipping init"
-    cd frappe-bench
-    bench start
-else
-    echo "Creating new bench..."
-fi
-
-bench init --skip-redis-config-generation frappe-bench
-
-cd frappe-bench
-
-# Use containers instead of localhost
-bench set-mariadb-host mariadb
-bench set-redis-cache-host redis:6379
-bench set-redis-queue-host redis:6379
-bench set-redis-socketio-host redis:6379
-
-# Remove redis, watch from Procfile
-sed -i '/redis/d' ./Procfile
-sed -i '/watch/d' ./Procfile
-
-bench get-app crm --branch develop
-
-bench new-site crm.localhost \
-    --force \
-    --mariadb-root-password 123 \
-    --admin-password admin \
-    --no-mariadb-socket
-
-bench --site crm.localhost install-app next_crm
-bench --site crm.localhost set-config developer_mode 1
-bench --site crm.localhost clear-cache
-bench --site crm.localhost set-config mute_emails 1
-bench use crm.localhost
-
-bench start
