+++ frontend/src/main.js
@@ -55,7 +55,7 @@ app.config.globalProperties.$dialog = createDialog
 
 let socket
 if (import.meta.env.DEV) {
-  frappeRequest({ url: '/api/method/next_crm.www.crm.get_context_for_dev' }).then(
+  frappeRequest({ url: '/api/method/next_crm.www.get_context_for_dev' }).then(
     (values) => {
       for (let key in values) {
         window[key] = values[key]
