+++ next_crm/www/__init__.py
@@ -1,0 +1,24 @@
+import frappe
+
+
+@frappe.whitelist(methods=["POST"], allow_guest=True)
+def get_context_for_dev():
+    if not frappe.conf.developer_mode:
+        frappe.throw("This method is only meant for developer mode")
+    return get_boot()
+
+
+def get_boot():
+    return frappe._dict(
+        {
+            "frappe_version": frappe.__version__,
+            "default_route": get_default_route(),
+            "site_name": frappe.local.site,
+            "read_only_mode": frappe.flags.read_only,
+            "csrf_token": frappe.sessions.get_csrf_token(),
+        }
+    )
+
+
+def get_default_route():
+    return "/next-crm"
