+++ next_crm/www/next-crm/index.py
@@ -4,6 +4,8 @@
 import frappe
 from frappe.utils.telemetry import capture
 
+from next_crm.www import get_boot
+
 no_cache = 1
 
 
@@ -14,26 +16,3 @@ def get_context():
     if frappe.session.user != "Guest":
         capture("active_site", "next-crm")
     return context
-
-
-@frappe.whitelist(methods=["POST"], allow_guest=True)
-def get_context_for_dev():
-    if not frappe.conf.developer_mode:
-        frappe.throw("This method is only meant for developer mode")
-    return get_boot()
-
-
-def get_boot():
-    return frappe._dict(
-        {
-            "frappe_version": frappe.__version__,
-            "default_route": get_default_route(),
-            "site_name": frappe.local.site,
-            "read_only_mode": frappe.flags.read_only,
-            "csrf_token": frappe.sessions.get_csrf_token(),
-        }
-    )
-
-
-def get_default_route():
-    return "/next-crm"
